Contact show component uses contact reducer
The rendering of a contact show component does not have the
same dependencies as contact form, such as investigation started at and
touched attributes, therefore it requires a separate reducer.

[#151921102]
