+++ app/javascript/selectors/contactFormSelectors.js
@@ -1,10 +1,4 @@
 import {createSelector} from 'reselect'
-import {
-  getStatusesSelector,
-  getPurposesSelector,
-  getLocationsSelector,
-  getCommunicationMethodsSelector,
-} from 'selectors/systemCodeSelectors'
 import {Map, fromJS, List} from 'immutable'
 import {dateTimeFormatter} from 'utils/dateFormatter'
 import {
@@ -15,31 +9,6 @@ import {
 } from 'utils/validator'
 import nameFormatter from 'utils/nameFormatter'
 
-/* eslint-disable no-invalid-this */
-const systemCodeDisplayValue = (code, systemCodes, noSetValue = Map(), context = this) => systemCodes.find(
-  (systemCode) => systemCode.get('code') === code, context, noSetValue
-).get('value')
-/* eslint-enable no-invalid-this */
-export const getStatusValueSelector = createSelector(
-  (state) => state.getIn(['contactForm', 'status', 'value']),
-  getStatusesSelector,
-  (code, systemCodes) => systemCodeDisplayValue(code, systemCodes)
-)
-export const getPurposeValueSelector = createSelector(
-  (state) => state.getIn(['contactForm', 'purpose', 'value']),
-  getPurposesSelector,
-  (code, systemCodes) => systemCodeDisplayValue(code, systemCodes)
-)
-export const getLocationValueSelector = createSelector(
-  (state) => state.getIn(['contactForm', 'location', 'value']),
-  getLocationsSelector,
-  (code, systemCodes) => systemCodeDisplayValue(code, systemCodes)
-)
-export const getCommunicationMethodValueSelector = createSelector(
-  (state) => state.getIn(['contactForm', 'communication_method', 'value']),
-  getCommunicationMethodsSelector,
-  (code, systemCodes) => systemCodeDisplayValue(code, systemCodes)
-)
 export const getPeopleSelector = createSelector(
   (state) => state.get('contactForm'),
   (contactForm) => contactForm.get('people') || List()
