+++ spec/javascripts/selectors/contactSelectorsSpec.js
@@ -1,0 +1,76 @@
+import {fromJS} from 'immutable'
+import {
+  getStatusValueSelector,
+  getPurposeValueSelector,
+  getLocationValueSelector,
+  getCommunicationMethodValueSelector,
+} from 'selectors/contactSelectors'
+import * as matchers from 'jasmine-immutable-matchers'
+
+describe('contactSelectors', () => {
+  beforeEach(() => jasmine.addMatchers(matchers))
+
+  describe('getStatusValueSelector', () => {
+    const contactStatuses = [{code: 'A', value: 'Attempted'}]
+
+    it('returns the current contact status display value', () => {
+      const contact = {status: 'A'}
+      const state = fromJS({contact, contactStatuses})
+      expect(getStatusValueSelector(state)).toEqual('Attempted')
+    })
+
+    it('returns undefined when contact does not have a status', () => {
+      const contact = {status: null}
+      const state = fromJS({contact, contactStatuses})
+      expect(getStatusValueSelector(state)).toEqual(undefined)
+    })
+  })
+
+  describe('getPurposeValueSelector', () => {
+    const contactPurposes = [{code: '123', value: 'SomePurpose'}]
+
+    it('returns the current contact purpose display value', () => {
+      const contact = {purpose: '123'}
+      const state = fromJS({contact, contactPurposes})
+      expect(getPurposeValueSelector(state)).toEqual('SomePurpose')
+    })
+
+    it('returns undefined when contact does not have a purpose', () => {
+      const contact = {purpose: null}
+      const state = fromJS({contact, contactPurposes})
+      expect(getPurposeValueSelector(state)).toEqual(undefined)
+    })
+  })
+
+  describe('getLocationValueSelector', () => {
+    const locations = [{code: '444', value: 'school'}]
+
+    it('returns the current contact location display value', () => {
+      const contact = {location: '444'}
+      const state = fromJS({contact, locations})
+      expect(getLocationValueSelector(state)).toEqual('school')
+    })
+
+    it('returns undefined when contact does not have a location', () => {
+      const contact = {location: null}
+      const state = fromJS({contact, locations})
+      expect(getLocationValueSelector(state)).toEqual(undefined)
+    })
+  })
+
+  describe('getCommunicationMethodValueSelector', () => {
+    const communicationMethods = [{code: '555', value: 'Phone'}]
+
+    it('returns the current contact communication methods display value', () => {
+      const contact = {communication_method: '555'}
+      const state = fromJS({contact, communicationMethods})
+      expect(getCommunicationMethodValueSelector(state)).toEqual('Phone')
+    })
+
+    it('returns undefined when contact does not have a communication method', () => {
+      const contact = {communication_method: null}
+      const state = fromJS({contact, communicationMethods})
+      expect(getCommunicationMethodValueSelector(state)).toEqual(undefined)
+    })
+  })
+})
